Moving compose e2e tests in their own files
